requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
Pillow>=10.0
psycopg[binary]>=3.1
gunicorn>=21.2
black>=24.0
isort>=5.13
